backend/requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.13.3
altgraph==0.17.4
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.64.0
anyio==3.7.1
APScheduler==3.10.4
argcomplete==3.6.2
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
blessed==1.21.0
blinker==1.9.0
boto3==1.39.12
botocore==1.39.12
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
claude-code==0.0.1
click==8.2.1
cohere==5.16.1
colorama==0.4.6
configparser==5.3.0
coverage==7.10.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.115.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastavro==1.11.1
feedparser==6.0.11
filelock==3.18.0
Flask==2.3.3
flask-cors==6.0.1
Flask-Pydantic==0.13.1
frozenlist==1.6.0
fsspec==2025.7.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==2.18.0
google-crc32c==1.7.1
google-genai==1.27.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
griffe==1.8.0
groq==0.30.0
gspread==5.12.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.10.0
jmespath==1.0.1
json_repair==0.47.8
jsonpath-python==1.0.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyboard==0.13.5
logfire-api==4.0.0
loguru==0.7.2
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.1
mdurl==0.1.2
mistralai==1.5.2
multidict==6.4.3
mutagen==1.47.0
mypy==1.11.2
mypy_extensions==1.1.0
numpy==2.2.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.97.1
openpyxl==3.1.5
opentelemetry-api==1.35.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pefile==2023.2.7
pillow==10.4.0
playwright==1.48.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.9.2
pydantic-ai==0.0.14
pydantic-ai-slim==0.0.14
pydantic-evals==0.4.6
pydantic-graph==0.4.6
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.19.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.2.1
pytest-mock==3.14.0
python-amazon-paapi==5.0.1
python-dateutil==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2024.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
ruff==0.7.0
s3transfer==0.13.1
selenium==4.27.1
sentry-sdk==2.34.1
setuptools==75.6.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sse-starlette==2.4.1
starlette==0.38.6
structlog==24.4.0
tenacity==8.5.0
tokenizers==0.21.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typer==0.16.0
types-beautifulsoup4==4.12.0.20250516
types-html5lib==1.1.11.20250708
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
watchfiles==1.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
windows-curses==2.4.1
wsproto==1.2.0
yarl==1.20.0
yt-dlp==2025.7.21
zipp==3.23.0
